Fix MyEventSystem Class Issue
Added a .asemref file and reference it to MyEventSystem Assembly in Gameanalytics>plugin (project)folder to solve the issue =>
MyEventSystem Class: MyEventSystem class is unable to find GameAnalytics required to send level start and finish events. Although the GameAnalytics unitypackage containing the script for handling events has been imported, it is not functioning as expected.